Add banking integration overall architecture diagram
Note: this diagram covers the banking integration work as a whole, not
only the CiviBanking extension. With CiviBanking forming the common
framework for the banking integration work, it seems the right place to
put such general documentation.

The diagram is stored as an uncompressed dia file. Hopefully this gives
somewhat more useful diffs than the (default) copressed files... Plus it
should make the repository smaller (as opposed to the working copy),
because this way delta compression can be applied between revisions.