Show whitespace sequences in UI if they're significant
In some parts of the UI where values may be used for searches
or as part of banking rules, whitespace sequences (i.e. two spaces)
may be significant and would cause a rule/search not to match when
these values are copied and pasted without the sequence.

HTML's default behaviour is to ignore whitespace sequences. This
changes the whitespace handling to show things like two consecutive
spaces so that values may be copied safely.

The change is applied to the purpose/detail view on the transaction
review page, and the purpose column of statement search.