Add HTML input types to entity definitions
This is required to make the fields available in FormBuilder, e.g. as
search filter.

Additionally the <add> element is removed which was meant to indicate
the CiviCRM version a field was added. (Makes no sense for an
extension.)

The line indent is now two spaces in all schema files.